# Optimize I/O operations in the TextFile utility class