This PR adds unit tests for `org.hl7.fhir.utilities.TextFile` and optimize I/O operations in it.

While analyzing the IG Publisher performance, I saw that methods in `TextFile` took a lot of time for simple I/O operations, so I optimized it.

For the performance changes, I measured building [CH Core](https://github.com/hl7ch/ch-core) before and after the patch:
|                             | Before         |                   | After          |                   |
|-----------------------------|----------------|-------------------|----------------|-------------------|
|                             | Execution time | Memory allocation | Execution time | Memory allocation |
| streamToString(InputStream) | 19.107 s       | 22.66 GB          | 0.140 s        | 300 MB            |
| fileToString(String)        | 5.355 s        | 6.01 GB           | 0.301 s        | 544 MB            |
| bytesToString(byte[])       | 3.527 s        | 4.19 GB           | 0.076 s        | 307 MB            |

The total build time went from 4min31 to 4min06 on those runs.

I also tested the IG Publisher on some other IGs, with success.

The BOM support for writing files is a bit surprising. I am unsure it is desired to write `.fhir/packages/packages.ini` without BOM, but the different _.index.json_ with BOM (maybe a method confusion?).
Renaming the methods could be helpful.

I wrote a small blog post that contains a bit more details: https://blog.qligier.ch/posts/2024-improving-perf-ig-publisher-io/