@grahamegrieve 2024-01-16
first of all, thanks for this. Great work. And I sure made a mess of the BOM thing. I refactored this in a separate code base to clarify the BOM confusion, but that means I can't easily accept this one directly. I'm going to migrate the actual optimisations to my branch, which is fine, that means you won't get recognised as a contributor directly :-(