@qligier 2024-01-16
> I'm going to migrate the actual optimisations to my branch

Great!

> that means you won't get recognised as a contributor directly :-(

That is no issue for me :) Ping me here or on Zulip if you need something more, I would be happy to help, and feel free to close this PR.