@grahamegrieve 2024-01-16
ok closed. All the code migrated to my other PR. You're always welcome to contribute efficiency fixes!